Move FeedbackUtil into separate package
